extract twitter handle module & improve tests
