+++ app/models/person.rb
@@ -1,26 +1,11 @@
 class Person < ActiveRecord::Base
+  include TwitterHandle
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :trackable, :validatable
   belongs_to :group
   mount_uploader :picture, PictureUploader
-
-  def twitter=(handle)
-    self[:twitter] = handle.gsub('@', '')
-  end
-
-  def has_twitter?
-    twitter && twitter != ''
-  end
-
-  def twitter_handle
-    '@' + twitter
-  end
-
-  def twitter_profile_url
-    'https://twitter.com/' + twitter
-  end
 
   def has_group?
     group && group != ''
